requirements.txt
Django==4.1.5
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.4.0